requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic-settings-2.7.1
pydantic[email]

asyncpg>=0.24.0

bcrypt==4.2.1 
passlib==1.7.4
pyjwt-2.10.1